+++ config/routes.rb
@@ -18,7 +18,15 @@ Rails.application.routes.draw do
     get :sitemap
   end
 
-  # Sleep Tips
+  # SEO Content
+  get 'about-sleep-apnea' => 'static#about_sleep_apnea'
+  get 'about-PAP-therapy' => 'static#about_PAP_therapy'
+  get 'PAP-setup-guide'   => 'static#PAP_setup_guide'
+  get 'PAP-troubleshooting-guide' => 'static#PAP_troubleshooting_guide'
+  get 'PAP-care-and-maintenance' => 'static#PAP_care_maintenance'
+  get 'PAP-masks-and-equipment' => 'static#PAP_masks_equipment'
+  get 'traveling-with-PAP' => 'static#traveling_with_PAP'
+  get 'side-effects-of-PAP' => 'static#side_effects_PAP'
   get 'sleep-tips'   => 'static#sleep_tips'
 
   # Registration flow