+++ test/mailers/user_mailer_test.rb
@@ -52,5 +52,19 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/Your registration with MyApnea\.Org was successful\. Thank you for joining our growing network!/, email.encoded)
   end
 
+  test "mentioned in post email" do
+    valid = users(:user_1)
+    post = posts(:one)
+
+    # Send the email, then test that it got queued
+    email = UserMailer.mentioned_in_post(post, valid).deliver_now
+    assert !ActionMailer::Base.deliveries.empty?
+
+    # Test the body of the sent email contains what we expect it to
+    assert_equal [valid.email], email.to
+    assert_equal "#{post.user.forum_name} Mentioned You on the MyApnea Forums", email.subject
+    assert_match(/#{post.user.forum_name} mentioned you in a post on the forums at MyApnea.Org\./, email.encoded)
+  end
+
 
 end